+++ packages/server/src/schema/thread/resolvers.ts
@@ -42,6 +42,7 @@ import {
 import { DiscriminatorType, UserDecoder, UuidDecoder } from '~/model/decoder'
 import { fetchScopeOfUuid } from '~/schema/authorization/utils'
 import { resolveConnection } from '~/schema/connection/utils'
+import { decodeSubjectId } from '~/schema/subject/utils'
 import { createUuidResolvers } from '~/schema/uuid/abstract-uuid/utils'
 import { Comment, Thread } from '~/types'
 
@@ -60,6 +61,9 @@ export const resolvers: InterfaceResolvers<'ThreadAware'> &
   },
   ThreadQuery: {
     async allThreads(_parent, input, { dataSources }) {
+      const subjectId = input.subjectId
+        ? decodeSubjectId(input.subjectId)
+        : null
       const limit = 50
       const { first = 10, instance } = input
       // TODO: Better solution
@@ -70,18 +74,12 @@ export const resolvers: InterfaceResolvers<'ThreadAware'> &
       if (first && first > limit)
         throw new UserInputError(`"first" cannot be larger than ${limit}`)
 
-      const { firstCommentIds } = await dataSources.model.serlo.getAllThreads({
-        first: first + 1,
-        after,
-        instance,
-      })
-
-      const threads = await resolveThreads({ firstCommentIds, dataSources })
-
-      // TODO: The types do not match
-      // TODO: Support for resolving small changes
       return resolveConnection({
-        nodes: threads,
+        nodes: await queryThreads({
+          first: first + 1,
+          threadsToFetch: first + 1,
+          after,
+        }),
         payload: { ...input, first, after },
         createCursor: (node) => {
           const comments = node.commentPayloads
@@ -90,6 +88,49 @@ export const resolvers: InterfaceResolvers<'ThreadAware'> &
           return latestComment.date
         },
       })
+
+      async function queryThreads({
+        first,
+        after,
+        threadsToFetch,
+        loopCount = 0,
+      }: {
+        threadsToFetch: number
+        first: number
+        after: string | undefined
+        loopCount?: number
+      }): Promise<Model<'Thread'>[]> {
+        const { firstCommentIds } = await dataSources.model.serlo.getAllThreads(
+          {
+            first: threadsToFetch,
+            after,
+            instance,
+          }
+        )
+
+        const threads = await resolveThreads({
+          firstCommentIds,
+          dataSources,
+          subjectId,
+        })
+
+        if (
+          threads.length < first &&
+          firstCommentIds.length === threadsToFetch &&
+          loopCount < 3
+        ) {
+          return threads.concat(
+            await queryThreads({
+              first: first - threads.length,
+              after: threads.at(-1)?.commentPayloads?.at(-1)?.date,
+              threadsToFetch,
+              loopCount: loopCount + 1,
+            })
+          )
+        } else {
+          return threads.slice(0, first)
+        }
+      }
     },
   },
   Thread: {